스나이퍼타워 추가, 데미지 로직 수정

+++ SBS_TowerDefense/Assets/Scripts/Managers/InterFaces.cs
@@ -26,7 +26,7 @@ public class InterFaces : MonoBehaviour
 
     public interface IAttackAble
     {
-        float attackPower { get; set; }
+        int attackPower { get; set; }
         float attackSpeed { get; set; }
         int attackRange { get; set; }
